jdk: Fix a whole bunch of warnings and a few build errors.
GCC is more strict than MSVC.

+++ trunk/openjdk/jdk/src/windows/native/sun/java2d/windows/Win32OffScreenSurfaceData.cpp
@@ -466,7 +466,7 @@ void Win32OSSD_GetRasInfo(JNIEnv *env,
     }
     if (wsdo->lockFlags & SD_LOCK_LUT) {
         pRasInfo->lutBase =
-            (long *) wsdo->device->GetSystemPaletteEntries();
+            (jint *) wsdo->device->GetSystemPaletteEntries();
         pRasInfo->lutSize = 256;
     } else {
         pRasInfo->lutBase = NULL;
